fix: keep text box anchor validation private
